+++ test/services/ldap-config/service.integration.test.js
@@ -151,6 +151,7 @@ describe('LdapConfigService', () => {
 			return {
 				setup: () => {},
 				getUsers: sinon.fake.resolves(fakeUsers),
+				verifyConfig: sinon.fake.resolves(fakeUsers),
 				getClasses: sinon.fake.resolves(fakeClasses),
 				disconnect: sinon.fake.resolves(),
 			};
@@ -223,7 +224,7 @@ describe('LdapConfigService', () => {
 		});
 
 		it('should catch common errors', async () => {
-			ldapServiceMock.getUsers = sinon.fake.rejects(new LDAPConnectionError());
+			ldapServiceMock.verifyConfig = sinon.fake.rejects(new LDAPConnectionError());
 			app.use('ldap', ldapServiceMock);
 
 			const result = await app
@@ -288,6 +289,7 @@ describe('LdapConfigService', () => {
 			return {
 				setup: () => {},
 				getUsers: sinon.fake.resolves(fakeUsers),
+				verifyConfig: sinon.fake.resolves(fakeUsers),
 				getClasses: sinon.fake.resolves(fakeClasses),
 				disconnect: sinon.fake.resolves(),
 			};